feat(Post): 좋아요 및 조회수 구현

+++ src/posts/service/post_service.py
@@ -16,7 +16,12 @@ class PostService:
     def get_post(self, post_id: int) -> Post:
         post = self.post_repository.find_by_id(post_id)
         if not post:
-            raise HTTPException(status_code=404, detail="Post not found")
+            raise HTTPException(status_code=404, detail="게시글을 찾을 수 없습니다.")
+
+        post.increase_view_count()
+
+        self.post_repository.save(post)
+
         return post
 
     def create_post(self, request: CreatePostRequest, member_id: int) -> Post:
@@ -29,10 +34,14 @@ class PostService:
         return self.post_repository.save(post)
 
     def update_post(self, post_id: int, request: UpdatePostRequest, member_id: int) -> Post:
-        post = self.get_post(post_id)
+
+        post = self.post_repository.find_by_id(post_id)
+
+        if not post:
+            raise HTTPException(status_code=404, detail="게시글을 찾을 수 없습니다.")
 
         if post.member_id != member_id:
-            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="수정 권한이 없습니다.")
 
         if request.title is not None:
             post.title = request.title
@@ -44,7 +53,13 @@ class PostService:
         return self.post_repository.save(post)
 
     def delete_post(self, post_id: int, member_id: int) -> None:
-        post = self.get_post(post_id)
+
+        post = self.post_repository.find_by_id(post_id)
+
+        if not post:
+            raise HTTPException(status_code=404, detail="게시글을 찾을 수 없습니다.")
+
         if post.member_id != member_id:
-            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="삭제 권한이 없습니다.")
+
         self.post_repository.delete(post)
